% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{build_network}
\alias{build_network}
\title{Build Networks from Coordinates}
\usage{
build_network(net, center_index, edge)
}
\arguments{
\item{net}{a data frame containing columns named as \strong{x_center} and \strong{y_center}, which are the x and y coordinates of cells, respectively.}

\item{center_index}{the index or name of the center cell.}

\item{edge}{the maximal length for an edge to exist.}
}
\value{
a network object.
}
\description{
This function is used to derive a network object from a coordinate matrix.
}
