% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralities.R
\name{centralities}
\alias{centralities}
\title{A Function Calculating Network Centralities Statistics}
\usage{
centralities(g, center_index)
}
\arguments{
\item{g}{A network object (g stands for "graph").}

\item{center_index}{The index or name of the vertex (center cell) whose centralities we are interested in.}
}
\description{
This function is used to calculate the network statistics with respect to one specific vertex \cr
in a network, most of which are various centralities. Reference: https://www.r-bloggers.com/2018/12/network-centrality-in-r-an-introduction/.
}
