% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_.R
\name{pca_}
\alias{pca_}
\title{Principal Component Analysis function used in the function quantile_matrix}
\usage{
pca_(data, pca_threshold = 0.9, ...)
}
\arguments{
\item{data}{the input data matrix, containing all the expression information like proteins or genes. \cr
Other things like indices and coordinates should not be included.}

\item{pca_threshold}{the minimal percentage of the explained variance in Principal Component Analysis. \cr
The default value is 0.9.}

\item{...}{other parameters could be passed to the \strong{prcomp} function.}
}
\description{
This function is used for Principal Component Analysis(PCA) in the \cr
function \strong{quantile_matrix} in this package. It uses \strong{prcomp} function for PCA.
}
